+++ frontend/compas_threejs/src/communications/messages.ts
@@ -11,12 +11,12 @@ import { themeManager } from "../viewer/theme_manager";
 import { uiManager } from "./sidebarStore";
 import { textManager } from "../viewer/text_manager";
 import { objectInfoManager } from "./objectInfo";
+import { objectActionManager } from "./objectInfo";
 import { removeObjectFromScene } from "../viewer/scene_manager";
 
 const SCENE_GEOMETRIES: { [guid: string]: THREE.Object3D } = {};
 
 export function dispatchMessage(message: Uint8Array) {
-    const messageUnpacked = unpackMessage(message);
     const obj = unpackMessageToGeometry(message);
 
     if (obj instanceof Dictionary) {
@@ -50,6 +50,10 @@ function analyzeDictionary(dictionary: Dictionary) {
             break;
         case "object_infos":
             objectInfoManager(data);
+            break;
+            return;
+        case "object_action":
+            objectActionManager(data);
             break;
             return;
         case "remove_object":