+++ frontend/compas_threejs/src/communications/objectInfo.ts
@@ -1,3 +1,5 @@
+import { objectActionsState } from "../store/store";
+import { sendData } from "@/communications/communication";
 import { objectBarData } from "../store/store";
 
 export function showObjectInfo() {
@@ -21,6 +23,33 @@ export function objectInfoManager(data: { [key: string]: any } | null) {
     if (data != null) {
         objectBarData.data = data;
     }
+}
+
+export function objectActionManager(data: Record<string, any>) {
+    console.log("Received object action:", data);
+    const action = {
+        guid: data.guid.value,
+        label: data.label.value,
+        type: data.type.value,
+        objectGuid: data.object_guid.value,
+        text: data.text.value,
+    };
+    objectActionsState.push(action);
+}
+
+export function handleObjectAction(action: { [key: string]: any }, value: any) {
+    const message = {
+        dispatch: "object_action_callback",
+        action_guid: action.guid,
+        value: null,
+        object_guid: action.objectGuid,
+    };
+
+    if (value !== undefined) {
+        message.value = value;
+    }
+
+    sendData(message);
 }
 
 document.addEventListener("keydown", (event) => {