+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 * Added, when an object is selected its material color changes to orange and slightly emissive yellow.
+* Added `actions` flag to `Viewer.add_geometry()` to add buttons and actions relative to one object.
+
 * Added `Viewer.update_metadata()` method.
 * Added, when an object is selected its material color changes to orange and slightly emissive yellow.
 * Added `Toolbar.vue` featuring existing transforms and views shortcuts into user-friendly buttons.